(chore)Renamed the report component
