+++ package.json
@@ -7,7 +7,7 @@
     "packages/*"
   ],
   "scripts": {
-    "start": "openmrs develop --backend https://ngx.ampath.or.ke --sources packages/esm-*-app  --api-url  /amrs --spa-path /amrs/spa/ --port 8040",
+    "start": "openmrs develop --backend https://ngx.ampath.or.ke --sources packages/esm-report-app  --api-url  /amrs --spa-path /amrs/spa/ --port 8040",
     "start:core": "openmrs develop --backend https://ngx.ampath.or.ke --sources packages/esm-ampath-core-app --api-url  /amrs --spa-path /amrs/spa/ --port 8030",
     "ci:publish": "yarn workspaces foreach --all --topological --exclude @ampath/esm-3.x-app npm publish --access public --tag latest",
     "ci:prepublish": "yarn workspaces foreach --all --topological --exclude @ampath/esm-3.x-app npm publish --access public --tag next",
