+++ packages/esm-report-app/src/reports-dashboard/left-justified-components/report-table-component.tsx
