+++ packages/esm-report-app/src/reports-dashboard/right-justified-components/date-pickers/date-pickers.scss
