+++ packages/esm-report-app/src/reports-dashboard/right-justified-components/generate-report-button.tsx
