+++ packages/esm-report-app/src/reports-dashboard/right-justified-components/styles/generate-report-component.scss
